fix notification types
