+++ Common/Types/NotificationRule/NotificationRuleType.ts
@@ -1,6 +1,7 @@
 enum NoticationRuleType {
-    INCIDENT_ASSIGNED = 'Incident Assigned',
-    ON_CALL_CHANGE = 'On Call Change',
+    ON_CALL_INCIDENT_CREATED = 'ON_CALL_INCIDENT_CREATED',
+    WHEN_USER_GOES_ON_CALL = 'When user goes on call',
+    WHEN_USER_GOES_OFF_CALL = 'When user goes off call',
 }
 
 export default NoticationRuleType;
