+++ Dashboard/src/Utils/PageMap.ts
@@ -88,7 +88,7 @@ enum PageMap {
     // User Settings.
     USER_SETTINGS = 'USER_SETTINGS',
     USER_SETTINGS_NOTIFICATION_METHODS = 'USER_SETTINGS_NOTIFICATION_METHODS',
-    USER_SETTINGS_NOTIFICATION_RULES = 'USER_SETTINGS_NOTIFICATION_RULES',
+    USER_SETTINGS_ON_CALL_RULES = 'USER_SETTINGS_ON_CALL_RULES',
 
     // Settings Routes
     SETTINGS = 'SETTINGS',
